refactor(web): flatten single-file directories and colocate tests
Closes LUM-2021

Flatten 9 single-file directories in apps/web/src/ and move 2 test
files out of __tests__/ to sit alongside their source files.

Directory flattening:
- components/terminal/ → components/terminal-panel.tsx
- components/apps/ → components/app-viewer-container.tsx
- domains/library/hooks/ → domains/library/use-library-data.ts
- domains/library/utils/ → domains/library/library-date.ts
- domains/account/components/ → domains/account/login-background.tsx
- domains/settings/components/devices/ → domains/settings/components/device-row.tsx
- domains/home/utils/ → domains/home/feed-utils.ts
- lib/environment/ → lib/environment-store.ts
- lib/errors/report.ts → lib/error-report.ts

Test colocation:
- tool-progress-card/__tests__/*.test.tsx → tool-progress-card/*.test.tsx

All import paths updated. Lint, typecheck, and affected tests pass.
