+++ apps/web/src/components/terminal-panel.tsx
