+++ apps/web/src/domains/chat/components/tool-progress-card/three-dot-indicator.test.tsx
