+++ apps/web/src/domains/chat/components/tool-progress-card/tool-progress-card-shell.test.tsx
