+++ apps/web/src/domains/settings/components/device-row.tsx
