+++ apps/web/src/lib/error-report.ts
